ListenerProducer/requirements.txt
websocket-client==0.57.0
kafka-python==1.4.7
bson==0.5.9
ksql==0.5.1.1
APScheduler==3.6.3
Flask==1.1.1
Flask-RESTful==0.3.8